requirements.txt
langchain
langchain-openai
openai
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.4
python-dotenv==1.0.0
